Safari-611.1.10/LayoutTests/media/track/video/video-track-mkv-theora-language.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../../media-file.js></script>
        <script src=../../video-test.js></script>
        <script src=../../in-band-tracks.js></script>
    </head>
    <body onload="testAttribute('../../content/two-audio-and-video-tracks.mkv', 'video', 'language', ['ru', 'zh'])">
        <video controls></video>
        <p>Check video tracks' language attributes.</p>
    </body>
</html>
